+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- Exceptions are emailed. The recipients and subject prefix can be configured.
+- Inline editing for grids.
 ### Changed
 ### Fixed