+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 - RMD form: a field can be set to submit the form on scan.
 - User permissions stored in `permission_tree` on users table.
 - EDI out process to update ERP system with Purchase Invoice prices handled by job queue.
+- Use the `Choices` js library instead of the `Selectr` library to make `select` tags searchable. This lib works better on mobile devices.
 ### Changed
 - Created RMD utilities route and moved check_barcode to it.
 ### Fixed